requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pyjwt
